requirements.txt
Django==4.2.11
mysqlclient
djangorestframework
requests
